// src/enums/LogTag.ts
export enum LogTag {
  D = 'L_DEBUG_O',
  A_INIT_O = 'A_INIT_O',
  A_SET_COMMAND_SWITCH_O = 'L_SET_COMMAND_SWITCH_O',
  A_REGISTER_CUSTOM_ELEMENT_O = 'A_REGISTER_CUSTOM_ELEMENT_O',
  L_REGISTER_DOM_ELEMENTS_O = 'L_REGISTER_DOM_ELEMENTS_O',
  A_REGISTER_MESSAGE_TYPE_O = 'A_REGISTER_MESSAGE_TYPE_O',
  A_SET_CUSTOM_SERVICE_O = 'A_SET_CUSTOM_SERVICE_O',
  A_SET_CUSTOM_LANGUAGE_PACK_O = 'A_SET_CUSTOM_LANGUAGE_PACK_O',
  A_CLONE_LANGUAGE_ENTRIES_O = 'A_CLONE_LANGUAGE_ENTRIES_O',
  A_REGISTER_LANGUAGE_PACK_O = 'A_REGISTER_LANGUAGE_PACK_O',
  A_SET_LANGUAGE_O = 'A_SET_LANGUAGE_O',
  A_READY_O = 'A_READY_O',
  L_REQ_USER_PROFILE_HOOK_T = 'L_REQ_USER_PROFILE_HOOK_T',
  L_REQ_USER_PROFILE_HOOK_R = 'L_REQ_USER_PROFILE_HOOK_R',
  L_REQ_USER_PROFILE_HOOK_E = 'L_REQ_USER_PROFILE_HOOK_E',
  L_GET_DEFAULT_USER_PROFILE_HOOK_E = 'L_GET_DEFAULT_USER_PROFILE_HOOK_E',
  L_GET_DEFAULT_USER_PROFILE_HOOK_O = 'L_GET_DEFAULT_USER_PROFILE_HOOK_O',
  A_UPDATE_USER_PROFILE_E = 'A_UPDATE_USER_PROFILE_E',
  A_UPDATE_USER_PROFILE_T = 'A_UPDATE_USER_PROFILE_T',
  A_UPDATE_USER_PROFILE_R = 'A_UPDATE_USER_PROFILE_R',
  A_UPDATE_USER_ONLINE_STATE_R = 'A_UPDATE_USER_ONLINE_STATE_R',
  A_UPDATE_USER_ONLINE_STATE_T = 'A_UPDATE_USER_ONLINE_STATE_T',
  L_GET_DEFAULT_GROUP_PROFILE_HOOK_O = 'L_GET_DEFAULT_GROUP_PROFILE_HOOK_O',
  L_GET_DEFAULT_GROUP_PROFILE_HOOK_E = 'L_GET_DEFAULT_GROUP_PROFILE_HOOK_E',
  L_REQ_GROUP_PROFILE_HOOK_T = 'L_REQ_GROUP_PROFILE_HOOK_T',
  L_REQ_GROUP_PROFILE_HOOK_E = 'L_REQ_GROUP_PROFILE_HOOK_E',
  L_REQ_GROUP_PROFILE_HOOK_R = 'L_REQ_GROUP_PROFILE_HOOK_R',
  A_UPDATE_GROUP_PROFILE_T = 'A_UPDATE_GROUP_PROFILE_T',
  A_UPDATE_GROUP_PROFILE_E = 'A_UPDATE_GROUP_PROFILE_E',
  A_UPDATE_GROUP_PROFILE_R = 'A_UPDATE_GROUP_PROFILE_R',
  L_REQ_GROUP_MEMBERS_HOOK_T = 'L_REQ_GROUP_MEMBERS_HOOK_T',
  L_REQ_GROUP_MEMBERS_HOOK_E = 'L_REQ_GROUP_MEMBERS_HOOK_E',
  L_REQ_GROUP_MEMBERS_HOOK_R = 'L_REQ_GROUP_MEMBERS_HOOK_R',
  A_UPDATE_GROUP_MEMBERS_T = 'A_UPDATE_GROUP_MEMBERS_T',
  A_UPDATE_GROUP_MEMBERS_E = 'A_UPDATE_GROUP_MEMBERS_E',
  A_UPDATE_GROUP_MEMBERS_R = 'A_UPDATE_GROUP_MEMBERS_R',
  A_ADD_GROUP_MEMBERS_T = 'A_ADD_GROUP_MEMBERS_T',
  A_ADD_GROUP_MEMBERS_E = 'A_ADD_GROUP_MEMBERS_E',
  A_ADD_GROUP_MEMBERS_R = 'A_ADD_GROUP_MEMBERS_R',
  A_REMOVE_GROUP_MEMBERS_T = 'A_REMOVE_GROUP_MEMBERS_T',
  A_REMOVE_GROUP_MEMBERS_E = 'A_REMOVE_GROUP_MEMBERS_E',
  A_REMOVE_GROUP_MEMBERS_R = 'A_REMOVE_GROUP_MEMBERS_R',
  L_REQ_USER_ONLINE_HOOK_T = 'L_REQ_USER_ONLINE_HOOK_T',
  L_REQ_USER_ONLINE_HOOK_E = 'L_REQ_USER_ONLINE_HOOK_E',
  L_REQ_USER_ONLINE_HOOK_R = 'L_REQ_USER_ONLINE_HOOK_R',
  L_REQ_SYSTEM_PROFILE_HOOK_T = 'L_REQ_SYSTEM_PROFILE_HOOK_T',
  L_REQ_SYSTEM_PROFILE_HOOK_E = 'L_REQ_SYSTEM_PROFILE_HOOK_E',
  L_REQ_SYSTEM_PROFILE_HOOK_R = 'L_REQ_SYSTEM_PROFILE_HOOK_R',
  L_GET_DEFAULT_SYSTEM_PROFILE_HOOK_O = 'L_GET_DEFAULT_SYSTEM_PROFILE_HOOK_O',
  L_GET_DEFAULT_SYSTEM_PROFILE_HOOK_E = 'L_GET_DEFAULT_SYSTEM_PROFILE_HOOK_E',
  L_GET_TAGS_FROM_LIB_E = 'L_GET_TAGS_FROM_LIB_E',
  A_ADD_UNVERIFIED_TAG_O = 'A_ADD_UNVERIFIED_TAG_O',
  A_ADD_UNVERIFIED_TAG_E = 'A_ADD_UNVERIFIED_TAG_E',
  A_ADD_TAG_E = 'A_ADD_TAG_E',
  A_UPDATE_TAG_E = 'A_UPDATE_TAG_E',
  A_REMOVE_TAG_E = 'A_REMOVE_TAG_E',
  A_SET_INPUT_MENU_E = 'A_SET_INPUT_MENU_E',
  A_SET_MESSAGE_BUBBLE_CFG_E = 'A_SET_MESSAGE_BUBBLE_CFG_E',
  A_UPDATE_USER_ONLINE_STATE_E = 'A_UPDATE_USER_ONLINE_STATE_E',
  A_OPEN_CONVERSATION_E = 'A_OPEN_CONVERSATION_E',
  A_SEND_MESSAGE_E = 'A_SEND_MESSAGE_E',
  A_INSERT_MESSAGE_E = 'A_INSERT_MESSAGE_E',
  A_SET_CONVERSATION_EXTENSION_E = 'A_SET_CONVERSATION_EXTENSION_E',
  L_PUSH_2_QUEUE_O = 'L_PUSH_2_QUEUE_O',
  L_TASK_QUEUE_EXECUTE_O = 'L_PICK_ITEM_FROM_QUEUE_O',
  L_SEND_MSG_T = 'L_SEND_MSG_T',
  L_SEND_MSG_R = 'L_SEND_MSG_R',
  L_UPLOAD_FILE_T = 'L_UPLOAD_FILE_T',
  L_UPLOAD_FILE_R = 'L_UPLOAD_FILE_R',
  L_UPLOAD_FILE_O = 'L_UPLOAD_FILE_O',
  A_SET_CHAT_EMOJI_LIBRARY_E = 'A_SET_CHAT_EMOJI_LIBRARY_E',
  L_GET_PUSH_CONFIG_HOOK_E = 'L_GET_PUSH_CONFIG_HOOK_E',
  L_SET_RECALL_DURATION_E = 'L_SET_RECALL_DURATION_E',
  L_SEND_READ_RECEIPT_MESSAGEV4_T = 'L_SEND_READ_RECEIPT_MESSAGEV4_T',
  L_SEND_READ_RECEIPT_MESSAGEV4_R = 'L_SEND_READ_RECEIPT_MESSAGEV4_R',
  L_PARSE_CONVERSATION_E = 'L_PARSE_CONVERSATION_E',
  L_GET_CONVERSATION_E = 'L_GET_CONVERSATION_E',
}